requirements.txt
python-dotenv
azure-identity
azure-ai-projects
azure-ai-inference
azure-search-documents
openai
semantic-kernel[azure] 